fixed bug security for get data

+++ server/src/main/java/fit/se/controllers/UserCtrl.java
@@ -3,6 +3,8 @@ package fit.se.controllers;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ExecutionException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,12 +34,37 @@ public class UserCtrl {
   @GetMapping(value = {
       "", "/"
   })
-  public List<User> getUsers() {
+  // bug: get all user but not return data for client
+  public ResponseEntity<ResponeMessage> getUsers() throws InterruptedException, ExecutionException {
     try {
+      List<Map<String, Object>> usersMap = new ArrayList<>();
       List<User> users = userService.getUsers();
-      return users;
+      for (User user : users) {
+        HashMap<String, Object> response = HashMapConverter.toHashMap(user);
+        response.remove("password");
+        usersMap.add(response);
+      }
+
+      return ResponseEntity.status(HttpStatus.OK).body(new ResponeMessage("ok", "success", usersMap));
     } catch (Exception e) {
-      return new ArrayList<>();
+      return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+  }
+
+  @PostMapping(value = {
+      "", "/add"
+  }, consumes = {
+      "application/json",
+      "application/x-www-form-urlencoded"
+  })
+  public ResponseEntity<ResponeMessage> addUser(@RequestBody User user) {
+    try {
+      userService.addUser(user);
+
+      return ResponseEntity.status(HttpStatus.OK).body(new ResponeMessage("ok", "success", null));
+    } catch (Exception e) {
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+          .body(new ResponeMessage("error", "Not found", e.getMessage()));
     }
   }
 
@@ -52,7 +80,7 @@ public class UserCtrl {
       return ResponseEntity.status(HttpStatus.OK).body(new ResponeMessage("ok", "success", response));
     } catch (Exception e) {
       return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-          .body(new ResponeMessage("error", "Not found", null));
+          .body(new ResponeMessage("error", "Not found", e.getMessage()));
     }
   }
 
@@ -62,12 +90,32 @@ public class UserCtrl {
       "application/json",
       "application/x-www-form-urlencoded"
   })
-  public void updateUser(@RequestBody User newUser, @PathVariable String userId) {
-    // userService.updateUser(newUser, userId);
+  public ResponseEntity<ResponeMessage> updateUser(@RequestBody User newUser) {
+    try {
+      User user = userService.getUser(newUser.getId());
+      if (user == null) {
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+      }
+      userService.updateUser(newUser);
+      return ResponseEntity.status(HttpStatus.OK).body(new ResponeMessage("ok", "success", null));
+    } catch (Exception e) {
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+          .body(new ResponeMessage("error", "Not found", e.getMessage()));
+    }
   }
 
   @DeleteMapping("/{userId}")
-  public void deleteUser(@PathVariable String userId) {
-    userService.deleteUser(userId);
+  public ResponseEntity<ResponeMessage> deleteUser(@PathVariable String userId) {
+    try {
+      User user = userService.getUser(userId);
+      if (user == null) {
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+      }
+      userService.deleteUser(userId);
+      return ResponseEntity.status(HttpStatus.OK).body(new ResponeMessage("ok", "success", null));
+    } catch (Exception e) {
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+          .body(new ResponeMessage("error", "Not found", e.getMessage()));
+    }
   }
 }
